+++ deploy_zsh.sh
@@ -1,4 +1,9 @@
 #!/bin/sh
-ln -s ./.zshrc $HOME/.zshrc
-ln -s ./.zprofile $HOME/.zprofile
+cd $HOME/dotfiles
+if [ ! -f $HOME/.zshrc ];then
+  ln -s ./.zshrc $HOME/.zshrc
+fi
+if [ ! -f $HOME/.zprofile ];then
+  ln -s ./.zprofile $HOME/.zprofile
+fi
 echo "created symbolic link to .zshrc, .zprofile"
